adds confirm dialog

+++ workers/data_refinery_workers/processors/management/commands/delete_local_files.py
@@ -1,3 +1,5 @@
+import sys
+
 from django.core.management.base import BaseCommand
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.models import (
@@ -10,11 +12,14 @@ logger = get_and_configure_logger(__name__)
 class Command(BaseCommand):
 
     def handle(self, *args, **options):
-        """ Deletes local original files
+        """ 
+        Deletes local original files
         """
-    for original_file in OriginalFile.objects.all():
-        try:
-            original_file.delete_local_file()
-        except Exception as e:
-            print("[" + str(original_file) + "]: " + str(e))
+        if not input("You're about to delete all of the local files! Proceed? (y/n): ").lower().strip()[0] == "y": sys.exit(1)
+
+        for original_file in OriginalFile.objects.all():
+            try:
+                original_file.delete_local_file()
+            except Exception as e:
+                print("[" + str(original_file) + "]: " + str(e))
